refactor: address all code review findings
HIGH:
- Extract normalizer types to normalizer-types.ts and tool helpers to
  tool-classify.ts, reducing normalizer.ts from 931 to 602 lines
- Fix /usr/local/bin/claude check gated by HOME in executor.ts
  (command.ts reverted to simple Bun.which — resolution is claude-specific)
- Fix duplicate step numbering in getBaseCommand comments

MEDIUM:
- Use top-level imports (node:fs, node:path) instead of require() in
  executor.ts for consistency with project conventions
- Capture `this` reference instead of `onResult` value in protocol.ts
  wrapStdout so callback can be set after stream wrapping
