+++ apps/api/src/engines/executors/claude/normalizer.ts
@@ -1,176 +1,31 @@
-import { classifyCommand } from '@/engines/logs'
-import type {
-  CommandCategory,
-  NormalizedLogEntry,
-  ToolAction,
-} from '@/engines/types'
+import type { NormalizedLogEntry } from '@/engines/types'
 import type { WriteFilterRule } from '@/engines/write-filter'
+import type {
+  ClaudeAssistant,
+  ClaudeContentItem,
+  ClaudeError,
+  ClaudeJson,
+  ClaudeRateLimit,
+  ClaudeResult,
+  ClaudeStreamEvent,
+  ClaudeSystem,
+  ClaudeToolResult,
+  ClaudeToolUse,
+  ClaudeUser,
+  ToolCallInfo,
+} from './normalizer-types'
+import {
+  buildToolResultRaw,
+  classifyToolAction,
+  classifyToolKind,
+  extractTextContent,
+  generateToolContent,
+  normalizeExecutionError,
+  normalizeToolResultContent,
+} from './tool-classify'
 
-// ---------- Claude JSON types (discriminated union) ----------
-
-/** Top-level message envelope from Claude CLI stdout (stream-json format). */
-type ClaudeJson =
-  | ClaudeSystem
-  | ClaudeAssistant
-  | ClaudeUser
-  | ClaudeToolUse
-  | ClaudeToolResult
-  | ClaudeStreamEvent
-  | ClaudeResult
-  | ClaudeError
-  | ClaudeRateLimit
-
-interface ClaudeSystem {
-  type: 'system'
-  subtype?: string
-  session_id?: string
-  cwd?: string
-  model?: string
-  tools?: unknown[]
-  apiKeySource?: string
-  status?: string
-  slash_commands?: string[]
-  plugins?: Array<{ name: string; path: string }>
-  agents?: string[]
-  compact_metadata?: Record<string, unknown>
-  output?: string
-  hook_name?: string
-  message?: string
-  content?: string
-  timestamp?: string
-}
-
-interface ClaudeAssistant {
-  type: 'assistant'
-  message: ClaudeMessage
-  session_id?: string
-  uuid?: string
-  timestamp?: string
-}
-
-interface ClaudeUser {
-  type: 'user'
-  message: ClaudeMessage
-  session_id?: string
-  uuid?: string
-  isSynthetic?: boolean
-  isReplay?: boolean
-  timestamp?: string
-}
-
-interface ClaudeToolUse {
-  type: 'tool_use'
-  id?: string
-  name?: string
-  input?: Record<string, unknown>
-  session_id?: string
-  timestamp?: string
-}
-
-interface ClaudeToolResult {
-  type: 'tool_result'
-  tool_use_id?: string
-  content?: string | unknown[]
-  is_error?: boolean
-  session_id?: string
-  timestamp?: string
-}
-
-interface ClaudeStreamEvent {
-  type:
-    | 'content_block_delta'
-    | 'content_block_start'
-    | 'content_block_stop'
-    | 'message_start'
-    | 'message_delta'
-    | 'message_stop'
-  index?: number
-  delta?: {
-    type?: string
-    text?: string
-    thinking?: string
-  }
-  content_block?: ClaudeContentItem
-  message?: ClaudeMessage
-  usage?: ClaudeUsage
-  parent_tool_use_id?: string
-  session_id?: string
-  uuid?: string
-  timestamp?: string
-}
-
-interface ClaudeResult {
-  type: 'result'
-  subtype?: string
-  is_error?: boolean
-  duration_ms?: number
-  cost_usd?: number
-  input_tokens?: number
-  output_tokens?: number
-  num_turns?: number
-  session_id?: string
-  result?: string
-  errors?: unknown[]
-  model_usage?: Record<string, { contextWindow?: number }>
-  usage?: ClaudeUsage
-  timestamp?: string
-}
-
-interface ClaudeError {
-  type: 'error'
-  error?: { type?: string; message?: string }
-  message?: string
-  timestamp?: string
-}
-
-interface ClaudeRateLimit {
-  type: 'rate_limit'
-  session_id?: string
-  rate_limit_info?: Record<string, unknown>
-  timestamp?: string
-}
-
-// ---------- Message / Content types ----------
-
-interface ClaudeMessage {
-  id?: string
-  role?: string
-  model?: string
-  content?: ClaudeContentItem[] | string
-  stop_reason?: string
-}
-
-type ClaudeContentItem =
-  | { type: 'text'; text: string }
-  | { type: 'thinking'; thinking: string }
-  | {
-      type: 'tool_use'
-      id?: string
-      name?: string
-      input?: Record<string, unknown>
-    }
-  | {
-      type: 'tool_result'
-      tool_use_id?: string
-      content?: string | unknown[]
-      is_error?: boolean
-    }
-
-interface ClaudeUsage {
-  input_tokens?: number
-  output_tokens?: number
-  cache_creation_input_tokens?: number
-  cache_read_input_tokens?: number
-  service_tier?: string
-}
-
-// ---------- Tool call info (for correlating tool_use → tool_result) ----------
-
-interface ToolCallInfo {
-  toolName: string
-  input: Record<string, unknown>
-  toolCallId: string
-}
+// Re-export for external consumers
+export { classifyToolAction, extractTextContent } from './tool-classify'
 
 // ---------- Normalizer ----------
 
@@ -743,189 +598,5 @@ export class ClaudeLogNormalizer {
     return this.rules.some(
       (r) => r.type === 'tool-name' && r.match === toolName,
     )
-  }
-}
-
-// ---------- Module-level helpers ----------
-
-function normalizeExecutionError(raw: string): {
-  kind?: string
-  summary: string
-} {
-  const lower = raw.toLowerCase()
-  if (
-    lower.includes('lsp server plugin:rust-analyzer-lsp') &&
-    lower.includes('crashed')
-  ) {
-    return {
-      kind: 'rust_analyzer_crash',
-      summary:
-        'Rust analyzer LSP crashed during execution. Retry the task or disable Rust tooling.',
-    }
-  }
-  const compact = raw.replace(/\s+/g, ' ').trim()
-  return {
-    summary: compact.length > 300 ? `${compact.slice(0, 300)}...` : compact,
-  }
-}
-
-/** Extract concatenated text from content blocks. */
-export function extractTextContent(content: unknown): string | null {
-  if (typeof content === 'string') return content
-  if (Array.isArray(content)) {
-    const texts = content
-      .filter((block: { type: string }) => block.type === 'text')
-      .map((block: { text: string }) => block.text)
-    return texts.length > 0 ? texts.join('') : null
-  }
-  return null
-}
-
-/** Generate concise, human-readable content for a tool invocation. */
-function generateToolContent(
-  toolName: string,
-  input: Record<string, unknown>,
-): string {
-  switch (toolName) {
-    case 'Read':
-      return String(input.file_path ?? input.path ?? toolName)
-    case 'Write':
-    case 'Edit':
-    case 'MultiEdit':
-      return String(input.file_path ?? input.path ?? toolName)
-    case 'Bash':
-      return String(input.command ?? input.cmd ?? toolName)
-    case 'Grep':
-      return input.path
-        ? `${input.pattern} in ${input.path}`
-        : String(input.pattern ?? toolName)
-    case 'Glob':
-      return input.path
-        ? `${input.pattern ?? input.filePattern} in ${input.path}`
-        : String(input.pattern ?? input.filePattern ?? toolName)
-    case 'LS':
-      return String(input.path ?? toolName)
-    case 'WebFetch':
-      return String(input.url ?? toolName)
-    case 'WebSearch':
-      return String(input.query ?? toolName)
-    case 'Task':
-      return input.description ? `Task: ${input.description}` : 'Task'
-    case 'TodoWrite':
-      return 'TODO list updated'
-    case 'ExitPlanMode':
-      return String(input.plan ?? 'Plan submitted')
-    case 'NotebookEdit':
-      return String(input.notebook_path ?? toolName)
-    default: {
-      // MCP tools: mcp__server__tool → mcp:server:tool
-      if (toolName.startsWith('mcp__')) {
-        const parts = toolName.split('__')
-        if (parts.length >= 3) {
-          return `mcp:${parts[1]}:${parts[2]}`
-        }
-      }
-      return `Tool: ${toolName}`
-    }
-  }
-}
-
-/** Classify tool kind for ToolDetail.kind field. */
-function classifyToolKind(toolName: string): string {
-  switch (toolName) {
-    case 'Read':
-      return 'file-read'
-    case 'Write':
-    case 'Edit':
-    case 'MultiEdit':
-      return 'file-edit'
-    case 'Bash':
-      return 'command-run'
-    case 'Grep':
-    case 'Glob':
-      return 'search'
-    case 'WebFetch':
-      return 'web-fetch'
-    case 'Task':
-      return 'task'
-    default:
-      return 'tool'
-  }
-}
-
-/** Classify a tool action (for ToolAction discriminated union). */
-export function classifyToolAction(
-  toolName: string,
-  input: Record<string, unknown>,
-): ToolAction {
-  switch (toolName) {
-    case 'Read':
-      return {
-        kind: 'file-read',
-        path: String(input.file_path ?? input.path ?? ''),
-      }
-    case 'Write':
-    case 'Edit':
-    case 'MultiEdit':
-      return {
-        kind: 'file-edit',
-        path: String(input.file_path ?? input.path ?? ''),
-      }
-    case 'Bash':
-      return {
-        kind: 'command-run',
-        command: String(input.command ?? input.cmd ?? ''),
-        category: classifyCommand(
-          String(input.command ?? input.cmd ?? ''),
-        ) as CommandCategory,
-      }
-    case 'Grep':
-    case 'Glob':
-      return {
-        kind: 'search',
-        query: String(input.pattern ?? input.query ?? input.filePattern ?? ''),
-      }
-    case 'WebFetch':
-      return { kind: 'web-fetch', url: String(input.url ?? '') }
-    default:
-      return { kind: 'tool', toolName, arguments: input }
-  }
-}
-
-/** Normalize tool_result content to a plain string. */
-function normalizeToolResultContent(
-  content: string | unknown[] | undefined,
-): string {
-  if (typeof content === 'string') return content
-  if (Array.isArray(content)) {
-    return content
-      .map((part: unknown) => {
-        if (typeof part === 'string') return part
-        if (
-          typeof part === 'object' &&
-          part !== null &&
-          'text' in part &&
-          typeof (part as { text: unknown }).text === 'string'
-        ) {
-          return (part as { text: string }).text
-        }
-        return JSON.stringify(part)
-      })
-      .join('\n')
-  }
-  return JSON.stringify(content ?? '')
-}
-
-/** Build raw object for tool result persistence. */
-function buildToolResultRaw(
-  info: ToolCallInfo,
-  resultContent: string,
-  isError?: boolean,
-): Record<string, unknown> {
-  return {
-    toolName: info.toolName,
-    input: info.input,
-    result: resultContent,
-    isError: isError ?? false,
   }
 }
